+++ twitter/server/lib/src/entities/tweet_entity/state.g.dart
@@ -12,12 +12,10 @@ TweetEntityState _$TweetEntityStateFromJson(Map<String, dynamic> json) =>
       (json['retweetedByUsers'] as List<dynamic>)
           .map((e) => e as String)
           .toList(),
-      DateTime.parse(json['createdAt'] as String),
     );
 
 Map<String, dynamic> _$TweetEntityStateToJson(TweetEntityState instance) =>
     <String, dynamic>{
-      'createdAt': instance.createdAt.toIso8601String(),
       'likedByUsers': instance._likedByUsers,
       'retweetedByUsers': instance._retweetedByUsers,
     };