+++ twitter/server/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Rename fields in view groups.
 - Use horda_server 0.13.0.
+- Reorganize business process code.
 
 ## 0.1.0